[#19] Add endpoint for exporting to unilog
Add export-instance endpoint. It expects a post with the instance name as part of the URL, checks if the instance can be exported and if so calls export-events in a future.

akvoreflow.unilogger/export-events is a wrapper that in turn calls unilogger/process-events which is used both in this way and by the gae endpoint.

Refactor unilogger/process-events to handle and save errors.

Add error_status and error_message columns to instance_status table to support process-events error logging.